requirements.txt
# Beancount - Dubbel bokföringssystem
beancount>=2.3.6

# Fava - Webbaserat gränssnitt för Beancount
fava>=1.27

# OCR och dokumentbehandling
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# AI/LLM Integration
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.20

# Databehandling
pandas>=2.0.0
python-dateutil>=2.8.2

# Konfiguration
python-dotenv>=1.0.0

# Validering och parsing
pydantic>=2.0.0

# Logging och debugging
colorlog>=6.7.0

# Kryptografi för OAuth
cryptography>=41.0.0
